# lib: Speed up reconnection attempts for zapi