Currently the zapi reconnection is once every 10 seconds for the first 3 times and then once every 60 seconds from then on out.  We are seeing interesting behavior under loaded systems where zebra is just slow to come up and daemons are spending a long time waiting to connect.  Let's just make things a bit more aggressive.

Change the code to attempt to reconnect once every second for 30 seconds and then change to once every 5 seconds from then on out.

This should help with non-integrated configuration on system startup.